refactor: broken up function test case down into separate classes and relocated them to the route package
set default app properties

rename 'indicator-mappings' to 'data-set-mappings'
