+++ src/main/resources/application.properties
@@ -1,6 +1,3 @@
-dhis2.api.username = admin
-dhis2.api.password = district
-
 spring.artemis.mode=EMBEDDED
 spring.artemis.host=localhost
 spring.artemis.port=61616
@@ -13,3 +10,7 @@ spring.datasource.initialization-mode=always
 spring.h2.console.enabled=true
 
 camel.dataformat.json-jackson.auto-discover-object-mapper=true
+
+management.endpoints.web.exposure.include=hawtio,jolokia
+spring.jmx.enabled=true
+hawtio.authenticationEnabled=false
