+++ src/test/java/org/hisp/dhis/integration/rapidpro/Environment.java
@@ -126,8 +126,8 @@ public final class Environment
             RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
 
             given( RAPIDPRO_REQUEST_SPEC ).contentType( ContentType.URLENC ).formParams(
-                    Map.of( "first_name", "Alice", "last_name", "Wonderland", "email", "claude@dhis2.org", "password",
-                        "12345678", "timezone", "Europe/Berlin", "name", "dhis2" ) )
+                Map.of( "first_name", "Alice", "last_name", "Wonderland", "email", "claude@dhis2.org", "password",
+                    "12345678", "timezone", "Europe/Berlin", "name", "dhis2" ) )
                 .when()
                 .post( "/org/signup/" ).then().statusCode( 302 );
 
@@ -144,7 +144,8 @@ public final class Environment
             ORG_UNIT_ID = createOrgUnit();
             createOrgUnitLevel();
             String orgUnitLevelId = null;
-            for ( OrganisationUnitLevel organisationUnitLevel : DHIS2_CLIENT.get( "organisationUnitLevels" ).withFields( "id" )
+            for ( OrganisationUnitLevel organisationUnitLevel : DHIS2_CLIENT.get( "organisationUnitLevels" )
+                .withFields( "id" )
                 .withoutPaging().transfer().returnAs( OrganisationUnitLevel.class, "organisationUnitLevels" ) )
             {
                 orgUnitLevelId = organisationUnitLevel.getId().get();
@@ -179,10 +180,10 @@ public final class Environment
 
         DHIS2_CONTAINER = new GenericContainer<>(
             "dhis2/core:2.36.10-tomcat-8.5.34-jre8-alpine" )
-            .withClasspathResourceMapping( "dhis.conf", "/DHIS2_home/dhis.conf", BindMode.READ_WRITE )
-            .withNetwork( DHIS2_NETWORK ).withExposedPorts( 8080 )
-            .waitingFor( new HttpWaitStrategy().forStatusCode( 200 ).withStartupTimeout( Duration.ofMinutes( 3 ) ) )
-            .withEnv( "WAIT_FOR_DB_CONTAINER", "db" + ":" + 5432 + " -t 0" );
+                .withClasspathResourceMapping( "dhis.conf", "/DHIS2_home/dhis.conf", BindMode.READ_WRITE )
+                .withNetwork( DHIS2_NETWORK ).withExposedPorts( 8080 )
+                .waitingFor( new HttpWaitStrategy().forStatusCode( 200 ).withStartupTimeout( Duration.ofMinutes( 3 ) ) )
+                .withEnv( "WAIT_FOR_DB_CONTAINER", "db" + ":" + 5432 + " -t 0" );
     }
 
     private static void composeRapidProContainers()
@@ -192,50 +193,50 @@ public final class Environment
 
         REDIS_CONTAINER = new GenericContainer<>(
             DockerImageName.parse( "redis:6.2.6-alpine" ) )
-            .withNetworkAliases( "redis" )
-            .withExposedPorts( 6379 )
-            .withNetwork( RAPIDPRO_NETWORK );
+                .withNetworkAliases( "redis" )
+                .withExposedPorts( 6379 )
+                .withNetwork( RAPIDPRO_NETWORK );
 
         ELASTICSEARCH_CONTAINER = new GenericContainer<>(
             DockerImageName.parse( "elasticsearch:6.8.23" ) )
-            .withEnv( "discovery.type", "single-node" )
-            .withNetwork( RAPIDPRO_NETWORK )
-            .withNetworkAliases( "elasticsearch" )
-            .withExposedPorts( 9200 )
-            .waitingFor( new HttpWaitStrategy().forStatusCode( 200 ) );
+                .withEnv( "discovery.type", "single-node" )
+                .withNetwork( RAPIDPRO_NETWORK )
+                .withNetworkAliases( "elasticsearch" )
+                .withExposedPorts( 9200 )
+                .waitingFor( new HttpWaitStrategy().forStatusCode( 200 ) );
 
         RAPIDPRO_CONTAINER = new GenericContainer<>(
             DockerImageName.parse( "praekeltfoundation/rapidpro:v7.2.4" ) )
-            .dependsOn( rapidProDbContainer )
-            .withExposedPorts( 8000 )
-            .withNetwork( RAPIDPRO_NETWORK )
-            .waitingFor( new HttpWaitStrategy().forStatusCode( 200 ).withStartupTimeout( Duration.ofMinutes( 3 ) ) )
-            .withEnv( "SECRET_KEY", "super-secret-key" )
-            .withEnv( "DATABASE_URL", "postgresql://temba:temba@db/temba" )
-            .withEnv( "REDIS_URL", "redis://redis:6379/0" )
-            .withEnv( "DJANGO_DEBUG", "on" )
-            .withEnv( "DOMAIN_NAME", "localhost" )
-            .withEnv( "MANAGEPY_COLLECTSTATIC", "on" )
-            .withEnv( "MANAGEPY_INIT_DB", "on" )
-            .withEnv( "MANAGEPY_MIGRATE", "on" )
-            .withEnv( "DJANGO_SUPERUSER_PASSWORD", "12345678" )
-            .withEnv( "MAILROOM_URL", "http://mailroom:8090" )
-            .withEnv( "MAILROOM_AUTH_TOKEN", "Gqcqvi2PGkoZMpQi" )
-            .withEnv( "ELASTICSEARCH_URL", "http://elasticsearch:9200" )
-            .withCommand( "sh", "-c",
-                "sed -i '/CsrfViewMiddleware/s/^/#/g' temba/settings_common.py && /startup.sh" );
+                .dependsOn( rapidProDbContainer )
+                .withExposedPorts( 8000 )
+                .withNetwork( RAPIDPRO_NETWORK )
+                .waitingFor( new HttpWaitStrategy().forStatusCode( 200 ).withStartupTimeout( Duration.ofMinutes( 3 ) ) )
+                .withEnv( "SECRET_KEY", "super-secret-key" )
+                .withEnv( "DATABASE_URL", "postgresql://temba:temba@db/temba" )
+                .withEnv( "REDIS_URL", "redis://redis:6379/0" )
+                .withEnv( "DJANGO_DEBUG", "on" )
+                .withEnv( "DOMAIN_NAME", "localhost" )
+                .withEnv( "MANAGEPY_COLLECTSTATIC", "on" )
+                .withEnv( "MANAGEPY_INIT_DB", "on" )
+                .withEnv( "MANAGEPY_MIGRATE", "on" )
+                .withEnv( "DJANGO_SUPERUSER_PASSWORD", "12345678" )
+                .withEnv( "MAILROOM_URL", "http://mailroom:8090" )
+                .withEnv( "MAILROOM_AUTH_TOKEN", "Gqcqvi2PGkoZMpQi" )
+                .withEnv( "ELASTICSEARCH_URL", "http://elasticsearch:9200" )
+                .withCommand( "sh", "-c",
+                    "sed -i '/CsrfViewMiddleware/s/^/#/g' temba/settings_common.py && /startup.sh" );
 
         MAILROOM_CONTAINER = new GenericContainer<>(
             DockerImageName.parse( "praekeltfoundation/mailroom:v7.0.1" ) )
-            .withNetwork( RAPIDPRO_NETWORK )
-            .withNetworkAliases( "mailroom" )
-            .withEnv( "MAILROOM_DOMAIN", "mailroom" )
-            .withEnv( "MAILROOM_ELASTIC", "http://elasticsearch:9200" )
-            .withEnv( "MAILROOM_ATTACHMENT_DOMAIN", "mailroom" )
-            .withEnv( "MAILROOM_AUTH_TOKEN", "Gqcqvi2PGkoZMpQi" )
-            .withEnv( "MAILROOM_DB", "postgres://temba:temba@db/temba?sslmode=disable" )
-            .withEnv( "MAILROOM_REDIS", "redis://redis:6379/0" )
-            .withCommand( "mailroom", "--address", "0.0.0.0" );
+                .withNetwork( RAPIDPRO_NETWORK )
+                .withNetworkAliases( "mailroom" )
+                .withEnv( "MAILROOM_DOMAIN", "mailroom" )
+                .withEnv( "MAILROOM_ELASTIC", "http://elasticsearch:9200" )
+                .withEnv( "MAILROOM_ATTACHMENT_DOMAIN", "mailroom" )
+                .withEnv( "MAILROOM_AUTH_TOKEN", "Gqcqvi2PGkoZMpQi" )
+                .withEnv( "MAILROOM_DB", "postgres://temba:temba@db/temba?sslmode=disable" )
+                .withEnv( "MAILROOM_REDIS", "redis://redis:6379/0" )
+                .withCommand( "mailroom", "--address", "0.0.0.0" );
     }
 
     private static PostgreSQLContainer<?> newPostgreSQLContainer( String databaseName,
@@ -243,19 +244,20 @@ public final class Environment
     {
         return new PostgreSQLContainer<>(
             DockerImageName.parse( "postgis/postgis:12-3.2-alpine" ).asCompatibleSubstituteFor( "postgres" ) )
-            .withDatabaseName( databaseName )
-            .withNetworkAliases( "db" )
-            .withUsername( username )
-            .withPassword( password ).withNetwork( network );
+                .withDatabaseName( databaseName )
+                .withNetworkAliases( "db" )
+                .withUsername( username )
+                .withPassword( password ).withNetwork( network );
     }
 
     private static void createIndicatorMappings()
         throws IOException
     {
-        DHIS2_CLIENT.post( "dataStore/rapidpro-connector/indicator-mappings" )
+        DHIS2_CLIENT.post( "dataStore/rapidpro-connector/data-set-mappings" )
             .withResource( StreamUtils.copyToString(
-                Thread.currentThread().getContextClassLoader().getResourceAsStream( "indicator-mappings.json" ),
-                Charset.defaultCharset() ) ).transfer();
+                Thread.currentThread().getContextClassLoader().getResourceAsStream( "data-set-mappings.json" ),
+                Charset.defaultCharset() ) )
+            .transfer();
     }
 
     private static void addOrgUnitToAdminUser( String orgUnitId )
@@ -307,18 +309,19 @@ public final class Environment
             .post( "/api/apitoken/refresh/" ).then().statusCode( 200 ).extract().path( "token" );
     }
 
-    private static void importMetaData(String orgUnitLevelId)
+    private static void importMetaData( String orgUnitLevelId )
         throws IOException
     {
         String metaData = StreamUtils.copyToString(
             Thread.currentThread().getContextClassLoader().getResourceAsStream( "MLAG00_1.2.1_DHIS2.36.json" ),
-            Charset.defaultCharset() ).replaceAll("<OU_LEVEL_DISTRICT_UID>", orgUnitLevelId).replaceAll( "<OU_LEVEL_FACILITY_UID>", orgUnitLevelId );
+            Charset.defaultCharset() ).replaceAll( "<OU_LEVEL_DISTRICT_UID>", orgUnitLevelId )
+            .replaceAll( "<OU_LEVEL_FACILITY_UID>", orgUnitLevelId );
 
         WebMessage webMessage = DHIS2_CLIENT.post( "metadata" )
             .withResource( metaData )
             .withParameter( "atomicMode", "NONE" ).transfer().returnAs( WebMessage.class );
 
-        assertEquals( DescriptiveWebMessage.Status.OK, webMessage.getStatus().get());
+        assertEquals( DescriptiveWebMessage.Status.OK, webMessage.getStatus().get() );
     }
 
     private static void createOrgUnitLevel()
@@ -326,13 +329,14 @@ public final class Environment
     {
         DHIS2_CLIENT.post( "filledOrganisationUnitLevels" )
             .withResource( Map.of( "organisationUnitLevels",
-                List.of( new OrganisationUnitLevel().withName( "Level 1" ).withLevel( 1 ) ) ) ).transfer().close();
+                List.of( new OrganisationUnitLevel().withName( "Level 1" ).withLevel( 1 ) ) ) )
+            .transfer().close();
     }
 
     private static String createOrgUnit()
     {
         return (String) ((Map<String, Object>) DHIS2_CLIENT.post( "organisationUnits" ).withResource(
-                new OrganisationUnit().withName( "Acme" ).withShortName( "Acme" ).withOpeningDate( new Date() ) ).transfer()
+            new OrganisationUnit().withName( "Acme" ).withShortName( "Acme" ).withOpeningDate( new Date() ) ).transfer()
             .returnAs( WebMessage.class ).getResponse().get()).get( "uid" );
     }
