+++ src/test/resources/application-test.properties
@@ -1,0 +1,2 @@
+dhis2.api.username = admin
+dhis2.api.password = district
